Rename text.* to menu.* and utils.* to utf8.* and move most function of main.c to menu.c
