--- utils/text.h
@@ -1,11 +1,0 @@
-#ifndef TEXT_H
-#define TEXT_H
-
-extern const char* const title[14];
-extern const int title_height;
-extern const int title_width;
-
-extern const int outer_menu_width;
-extern const int inner_menu_width;
-
-#endif
